finished perlin fr fr

+++ src/perlin.rs
@@ -49,9 +49,9 @@ impl Perlin {
     }
     pub fn noise(&self, p: &Point3) -> f64 {
         // Don't fully understand the mechanics here with the &
-        let i = (4. * p.x()) as usize & 255;
-        let j = (4. * p.y()) as usize & 255;
-        let k = (4. * p.z()) as usize & 255;
+        let i = ((4. * p.x()) as i32 & 255) as usize;
+        let j = ((4. * p.y()) as i32 & 255) as usize;
+        let k = ((4. * p.z()) as i32 & 255) as usize;
 
         return self.randfloat[(self.perm_x[i] ^ self.perm_y[j] ^ self.perm_z[k]) as usize];
     }
